URBNG-298 - [WP] Create database structure for tasks + small changes + version's type changed from double to int + fixed problem with TaskSolution entity

+++ UrbanGameWP7/Common/Tasks/IBaseTask.cs
@@ -32,6 +32,6 @@ namespace Common
 
         DateTime? EndDate { get; set; }
 
-        double Version { get; set; }
+        int Version { get; set; }
     }
 }
